requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
gunicorn==21.2.0 